LayoutTests/fast/events/ios/key-command-transpose-expected.txt

Tests that pressing Control + t in a content-editable field performs a tranpose.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


To run this test by hand, place the text insertion point before the 'h' in the text below and press Control + t.

PASS document.getElementById("test").textContent is "ha"
PASS successfullyParsed is true

TEST COMPLETE
